requirements.txt
atlassian-python-api
pydantic
python-dotenv
requests
